Make Accordion as separate component from NetworkTabs
