+++ projects/extension/src/components/NetworkTab.tsx
@@ -1,13 +1,11 @@
 import React, {
-  createContext,
   FunctionComponent,
   ReactNode,
   useCallback,
-  useContext,
   useMemo,
   useState,
 } from "react"
-import { StatusCircle } from "."
+import { Accordion, StatusCircle } from "."
 
 import { NetworkTabProps, App, OptionsNetworkTabHealthContent } from "../types"
 import "../main.css"
@@ -35,38 +33,6 @@ interface NetworkContentProps {
   health: OptionsNetworkTabHealthContent
   apps: App[]
   network: string
-}
-
-interface AccordionProps {
-  titles: ReactNode[] | string[]
-  titleClass?: string
-  contents: ReactNode[] | string[]
-  contentClass?: string
-  defaultExpanded?: number
-}
-
-interface AccItem {
-  value: string
-  title: ReactNode
-  children: ReactNode
-  titleClass?: string
-  contentClass?: string
-  status?: "first" | "last" | "single"
-}
-
-const AccordionContext = createContext({
-  activeItem: "",
-  setToggle: (val: string) => {
-    console.log(val)
-  },
-})
-
-const useAccordionContext = () => {
-  const context = useContext(AccordionContext)
-  if (!context) {
-    throw new Error("No context found for Accordion")
-  }
-  return context
 }
 
 const NetworkContent = ({ network, health, apps }: NetworkContentProps) => {
@@ -106,127 +72,6 @@ const NetworkContent = ({ network, health, apps }: NetworkContentProps) => {
         <div className="basis-2/3"></div>
       </div>
     </div>
-  )
-}
-
-const AccordionItem = ({
-  value,
-  title,
-  children,
-  titleClass,
-  contentClass,
-  status,
-}: AccItem) => {
-  const { activeItem, setToggle } = useAccordionContext()
-
-  const expanded = activeItem === value
-
-  return (
-    <div className={status && `item _mi__${status}`}>
-      <button
-        className={`item_title `
-          .concat(
-            `${
-              expanded && (status === "single" || status === "last")
-                ? "content_closed "
-                : ""
-            }`,
-          )
-          .concat(titleClass || "")}
-        aria-controls={`${value}-panel`}
-        aria-disabled="false"
-        aria-expanded={expanded}
-        id={`${value}-header`}
-        onClick={() => setToggle(value)}
-        type="button"
-        value={value}
-      >
-        {title}
-      </button>
-      <section
-        className={`item_content `
-          .concat(
-            `${
-              expanded && (status === "single" || status === "last")
-                ? "content_expanded "
-                : ""
-            }`,
-          )
-          .concat(contentClass || "")}
-        aria-hidden={!expanded}
-        aria-labelledby={`${value}-header`}
-        hidden={!expanded}
-      >
-        {children}
-      </section>
-    </div>
-  )
-}
-
-const Accordion = ({
-  titles,
-  contents,
-  defaultExpanded,
-  titleClass,
-  contentClass,
-}: AccordionProps): JSX.Element => {
-  const [activeItem, setActiveItem] = useState<string | undefined>(
-    defaultExpanded?.toString(),
-  )
-  const setToggle = useCallback(
-    (value: string) => {
-      setActiveItem(() => {
-        if (activeItem !== value) return value
-        return ""
-      })
-    },
-    [setActiveItem, activeItem],
-  )
-
-  const value = useMemo(
-    () => ({
-      activeItem,
-      setToggle,
-      defaultExpanded,
-    }),
-    [setToggle, activeItem, defaultExpanded],
-  )
-
-  if (titles.length !== contents.length) {
-    console.error("Titles do not have same length as contents.")
-    return (
-      <div>
-        Titles given do not have same length as contents. Please check the input
-        parameters.
-      </div>
-    )
-  }
-
-  return (
-    // @ts-ignore
-    <AccordionContext.Provider value={value}>
-      {titles.map((title, index) => {
-        const stat =
-          titles.length === 1
-            ? "single"
-            : index === 0
-            ? "first"
-            : index === titles?.length - 1
-            ? "last"
-            : undefined
-        return (
-          <AccordionItem
-            title={title}
-            value={index.toString()}
-            titleClass={titleClass}
-            contentClass={contentClass}
-            status={stat}
-          >
-            {contents[index]}
-          </AccordionItem>
-        )
-      })}
-    </AccordionContext.Provider>
   )
 }
