+++ projects/extension/src/components/index.tsx
@@ -5,3 +5,4 @@ export { default as ConnectedTab } from "./ConnectedTab"
 export { default as NetworkTab } from "./NetworkTab"
 export { default as Loader } from "./Loader"
 export { Tabs } from "./Tabs"
+export { Accordion } from "./Accordion"
